k8s-extension-prototype/controller/requirements.txt
PyYAML>=6.0
pandas>=2.0
gurobipy>=13.0
